% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map the state in the provided year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A single integer value representing the state code}

\item{year}{A single positive integer value representing the year of interest}
}
\description{
Draws a map of the specified state showing the reported accidents coordinates for the given year
}
\note{
Throws an error if the state number isn't valid or if there are no accidents to plot. Also doesn't work if provided parameters are vectors.
}
\examples{
\dontrun{
fars_map_state(1, 2013)

# Doesn't work
fars_map_state(c(1,2), c(2013,2014))
fars_map_state(c(1,2), 2013)
}
}
