% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read Fatality Analysis Reporting System file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Path to the file to read}
}
\value{
a tibble containing the data of interest
}
\description{
Imports the specified file and returns the data as a tibble object.
}
\note{
Results in an error if the provided file doesn't exist. Only a single string can be provided at once.
}
\examples{
\dontrun{
df <- fars_read("accident_2013.csv.bz2")

# this doesn't work
df <- fars_read(c("accident_2013.csv.bz2","accident_2014.csv.bz2"))
}
}
