% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read Fatality Analysis Reporting System file by providing the year}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of positive integers representing years of interest}
}
\value{
A list of tibbles
}
\description{
Imports the specified file by only providing the year as a parameter and returns a data frame containing
only MONTH and year columns.
}
\note{
Returns an error when supplying the function with years for which there are no data frames to import
}
\examples{
\dontrun{
rtbls <- fars_read_years(2014)
rtbls <- fars_read_years(c(2013,2014,2015))
}
}
