% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Count the number of entries per year and month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of positive integers representing years of interest}
}
\value{
A tibble containing 1 row and a number n of columns corresponding to the number of years provided as a parameter
}
\description{
Counts the number of entries grouped by year and month and returns the result as a tibble
}
\examples{
\dontrun{
df <- fars_summarize_years(2014)
df <- fars_summarize_years(c(2013,2014,2015))
}
}
